// src/app/models/school-admin/school-room-data-view.ts
/**
 * SchoolRoomDataView
 */
export interface SchoolRoomDataView { 
    // /**
    //  * Gets or sets SchoolFloorId
    //  */
    // schoolFloorId: string;
    /**
     * Gets or sets RoomNumber
     */
    roomNumber?: string;
    /**
     * Gets or sets Name
     */
    name: string;
    /**
     * Gets or sets RoomCapacity
     */
    roomCapacity?: number;
    /**
     * Gets or sets TypeId
     */
    typeId: number;
    /**
     * Gets or sets TypeId
     */
    id?: string;
}