// src/app/models/school-admin/school-building-update-view.ts
/**
 * SchoolbuildingUpdateView
 */
export interface SchoolBuildingUpdateView { 
    /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or Sets Name
     */
    name: string;
    /**
     * Gets or Sets Code
     */
    code?: string;
    /**
     * Gets or Sets SchoolId
     */
    schoolId?: string;
}